# Dictionary optimization after deserialization