@JeffreyZhao 2019-08-09
<details>
<summary>Benchmark code</summary>

```csharp
public class DictionaryPerf
{
    private Dictionary<int, int> _original;
    private Dictionary<int, int> _deserialized;

    public DictionaryPerf()
    {
        _original = new Dictionary<int, int> {
            {0, 0}, {1, 1}, {2, 2}, {3, 3}, {4, 4},
            {5, 5}, {6, 6}, {7, 7}, {8, 8}, {9, 9}
        };

        var formatter = new BinaryFormatter();
        var stream = new MemoryStream();
        formatter.Serialize(stream, _original);

        stream.Position = 0;
        _deserialized = (Dictionary<int, int>)formatter.Deserialize(stream);
    }

    [Benchmark(Baseline = true, OperationsPerInvoke = 10)]
    public void Original()
    {
       _original.ContainsKey(0); 
       _original.ContainsKey(1); 
       _original.ContainsKey(2); 
       _original.ContainsKey(3); 
       _original.ContainsKey(4); 
       _original.ContainsKey(5); 
       _original.ContainsKey(6); 
       _original.ContainsKey(7); 
       _original.ContainsKey(8); 
       _original.ContainsKey(9); 
    }

    [Benchmark(OperationsPerInvoke = 10)]
    public void Deserialized()
    {
       _deserialized.ContainsKey(0); 
       _deserialized.ContainsKey(1); 
       _deserialized.ContainsKey(2); 
       _deserialized.ContainsKey(3); 
       _deserialized.ContainsKey(4); 
       _deserialized.ContainsKey(5); 
       _deserialized.ContainsKey(6); 
       _deserialized.ContainsKey(7); 
       _deserialized.ContainsKey(8); 
       _deserialized.ContainsKey(9); 
    }
}
```

</details>

Before the fix:

```ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17763.615 (1809/October2018Update/Redstone5)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview7-012821
  [Host]     : .NET Core 3.0.0-preview7-27912-14 (CoreCLR 4.700.19.32702, CoreFX 4.700.19.36209), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview7-27912-14 (CoreCLR 4.700.19.32702, CoreFX 4.700.19.36209), 64bit RyuJIT
```

|       Method |     Mean |     Error |    StdDev | Ratio |
|------------- |---------:|----------:|----------:|------:|
|     Original | 6.383 ns | 0.0395 ns | 0.0330 ns |  1.00 |
| Deserialized | 9.022 ns | 0.0596 ns | 0.0557 ns |  1.41 |

After fix:

|       Method |     Mean |     Error |    StdDev | Ratio |
|------------- |---------:|----------:|----------:|------:|
|     Original | 6.631 ns | 0.0447 ns | 0.0419 ns |  1.00 |
| Deserialized | 6.633 ns | 0.0437 ns | 0.0409 ns |  1.00 |